Better Spider Movement, Fixed Builds Not Rendering
Spider moves in a zig-zag pattern.
Fixed builds not rendering graphics.
Dividing with * when getting half of values.

+++ RetroGame/Sources/PopUpText.cpp
@@ -24,8 +24,8 @@ void PopUpText::Draw(GameState* a_pGameState)
 {
 	if (a_pGameState)
 	{
-		a_pGameState->GetManager()->DrawString(m_position.GetX(),
-			m_position.GetY(),
+		a_pGameState->GetManager()->DrawString((int32_t)m_position.GetX(),
+			(int32_t)m_position.GetY(),
 			m_text,
 			olc::WHITE);
 	}
@@ -40,7 +40,7 @@ void PopUpText::Update(float a_fDeltaTime)
 	else
 	{
 		m_fDisplayLength -= a_fDeltaTime;
-		m_position -= m_position.Up() * m_uiDriftSpeed * a_fDeltaTime;
+		m_position -= m_position.Up() * (float)m_uiDriftSpeed * a_fDeltaTime;
 	}
 }
